CHROMIUM: platform/x86: chromeos_acpi: cast sizeof(array) to u32 before use
Removes a compiler warning:
"comparison of distinct pointer types lacks a cast"

BUG=chromium-os:5542
TEST=emerge chromeos-kernel, should not see warnings in chromeos_acpi.o

Reviewed-on: https://gerrit.chromium.org/gerrit/41483

+++ drivers/platform/x86/chromeos_acpi.c
@@ -442,7 +442,7 @@ static void handle_nested_acpi_package(union acpi_object *po, char *pm,
 
 		case ACPI_TYPE_STRING:
 			copy_size = min(element->string.length,
-					sizeof(attr_value) - 1);
+					(u32)(sizeof(attr_value)) - 1);
 			memcpy(attr_value, element->string.pointer, copy_size);
 			attr_value[copy_size] = '\0';
 			paa = create_sysfs_attribute(attr_value, pm, count, i);
@@ -621,7 +621,7 @@ static void handle_acpi_package(union acpi_object *po, char *pm)
 
 		case ACPI_TYPE_STRING:
 			copy_size = min(element->string.length,
-					sizeof(attr_value) - 1);
+					(u32)(sizeof(attr_value)) - 1);
 			memcpy(attr_value, element->string.pointer, copy_size);
 			attr_value[copy_size] = '\0';
 			add_sysfs_attribute(attr_value, pm, count, j);
